Фильтра сделаны, оформление пагинации

+++ app/Http/Controllers/Admin/ProductController.php
@@ -4,8 +4,8 @@ namespace App\Http\Controllers\Admin;
 
 use App\Category;
 use App\Http\Controllers\Controller;
+use App\Http\Requests\ProductRequest;
 use App\Product;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
 class ProductController extends Controller
@@ -17,7 +17,7 @@ class ProductController extends Controller
      */
     public function index()
     {
-        $products = Product::get();
+        $products = Product::paginate(10);
         return view('auth.products.index', compact('products'));
     }
 
@@ -40,9 +40,14 @@ class ProductController extends Controller
      */
     public function store(ProductRequest $request)
     {
-        $path = $request->file('image')->store('products');
         $params = $request->all();
-        $params['image'] = $path;
+
+        unset($params['image']);
+        if($request->has('image')){
+            $params['image'] = $request->file('image')->store('products');
+        }
+
+
         Product::create($params);
         return redirect()->route('products.index');
     }
@@ -79,10 +84,21 @@ class ProductController extends Controller
      */
     public function update(ProductRequest $request, Product $product)
     {
-        Storage::delete($product->image);
-        $path = $request->file('image')->store('products');
+
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+
+        if($request->has('image')){
+            Storage::delete($product->image);
+            $params['image'] = $request->file('image')->store('products');
+        }
+
+        foreach (['new', 'hit', 'recommend'] as $fieldName){
+            if(!isset($params[$fieldName])){
+                $params[$fieldName] = 0;
+            }
+        }
+
         $product->update($params);
         return redirect()->route('products.index');
     }
